:art: NamentlicheAnforderung into component
